Attendance Management/requirements.txt
click==8.1.3
cmake==3.22.5
colorama==0.4.4
dlib @ file:///C:/Users/hanan/Downloads/dlib-19.22.99-cp39-cp39-win_amd64.whl
face-recognition==1.3.0
face-recognition-models==0.3.0
mysql-connector-python==8.0.29
numpy==1.22.4
opencv-python==4.6.0.66
Pillow==9.1.1
protobuf==4.21.1
